requirements.txt
alembic==1.13.3
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
fastapi==0.115.0
uvicorn==0.31.0
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
jinja2==3.1.4